Google Test now downloaded as a part of the build system.
Many references point this to be the best way to use gtest:

- http://stackoverflow.com/questions/9689183/cmake-googletest
- http://www.kaizou.org/2014/11/gtest-cmake/
- https://gist.github.com/oneamtu/3734295

This changes are an adaptation of the most common technics to also use
the github repository (which is now the official source), and some
simpler cross-platform handling, though there was no test done in
Windows which has yet to be done.

The project should now success on Travis.